Django-DBs/Django-MongoDB/app/requirements.txt
Django==3.2.6
djongo==1.3.6
django-cors-headers==3.4.0
djangorestframework==3.12